requirements.txt
fastapi
uvicorn
streamlit
langchain
langchain-community
langchain-chroma
langchain-huggingface
chromadb
sentence-transformers
ddgs
groq
pydantic
Pillow
python-dotenv